+++ packages/eslint-plugin-internal/src/rules/no-typescript-default-import.ts
@@ -21,7 +21,6 @@ export default createRule({
     docs: {
       description:
         "Enforce that packages rules don't do `import ts from 'typescript';`",
-      recommended: 'recommended',
     },
     fixable: 'code',
     schema: [],