+++ packages/eslint-plugin/src/configs/stylistic-type-checked.ts
@@ -23,7 +23,6 @@ export = {
     '@typescript-eslint/no-confusing-non-null-assertion': 'error',
     'no-empty-function': 'off',
     '@typescript-eslint/no-empty-function': 'error',
-    '@typescript-eslint/no-empty-interface': 'error',
     '@typescript-eslint/no-inferrable-types': 'error',
     '@typescript-eslint/non-nullable-type-assertion-style': 'error',
     '@typescript-eslint/prefer-for-of': 'error',