+++ packages/eslint-plugin/src/rules/no-unused-vars.ts
@@ -104,7 +104,7 @@ export default createRule<Options, MessageIds>({
         vars: 'all',
         args: 'after-used',
         ignoreRestSiblings: false,
-        caughtErrors: 'none',
+        caughtErrors: 'all',
       };
 
       if (typeof firstOption === 'string') {
@@ -245,9 +245,7 @@ export default createRule<Options, MessageIds>({
           ) {
             continue;
           }
-        }
-
-        if (def.type === TSESLint.Scope.DefinitionType.Parameter) {
+        } else if (def.type === TSESLint.Scope.DefinitionType.Parameter) {
           // if "args" option is "none", skip any parameter
           if (options.args === 'none') {
             continue;
@@ -419,10 +417,23 @@ export default createRule<Options, MessageIds>({
                 ref.from.variableScope === unusedVar.scope.variableScope,
             );
 
+            const id = writeReferences.length
+              ? writeReferences[writeReferences.length - 1].identifier
+              : unusedVar.identifiers[0];
+
+            const { start } = id.loc;
+            const idLength = id.name.length;
+
+            const loc = {
+              start,
+              end: {
+                line: start.line,
+                column: start.column + idLength,
+              },
+            };
+
             context.report({
-              node: writeReferences.length
-                ? writeReferences[writeReferences.length - 1].identifier
-                : unusedVar.identifiers[0],
+              loc,
               messageId: 'unusedVar',
               data: unusedVar.references.some(ref => ref.isWrite())
                 ? getAssignedMessageData(unusedVar)