+++ packages/typescript-eslint/src/configs/all.ts
@@ -63,7 +63,7 @@ export default (
       '@typescript-eslint/no-dynamic-delete': 'error',
       'no-empty-function': 'off',
       '@typescript-eslint/no-empty-function': 'error',
-      '@typescript-eslint/no-empty-interface': 'error',
+      '@typescript-eslint/no-empty-object-type': 'error',
       '@typescript-eslint/no-explicit-any': 'error',
       '@typescript-eslint/no-extra-non-null-assertion': 'error',
       '@typescript-eslint/no-extraneous-class': 'error',